Use direct event listeners instead of event delegation
Apply click and transition end event listeners directly to the elements that
need them instead of applying them to the document and relying on event
delegation. We also now use vanilla JS to add and remove event listeners instead
of jQuery, and we remove the event listeners if the layouts are removed the
page.