feat: Improve navbar compatiblity for Bootstrap 5 and light/dark modes (#1145)
* fix(navbar): Don't flip navbar bg in dark mode

This rule only changes the background color of the navbar,
which is not enough to fully restyle the navbar in dark mode.
It might work in the default cases, but it certainly does not
work if users have set $navbar-dark-bg or $navbar-light-bg
to appropriate values. Note that Bootstrap expects the navbar
colors to be static (i.e. the same in light and dark mode), so if
we want to support different navbar colors we'll need to
implement something better.

* fix(bootstrap): Navbar toggler icon should follow navbar color mode

Bootstrap has the navbar toggler icon follow the global color mode, which makes the icon invisible when placed on a light background in dark mode. Outside of this rule, Bootstrap expects the navbar color to be consistent in light and dark mode.

* fix(bootstrap): Keep global color mode follow

* fix: Fix `_navbar.scss` patch

* feat: Only use `.navbar-default` and `.navbar-inverse` classes in BS <5

* fix(navbar_compat): navbar light background

* feat: account for navbar in light mode region

* feat: use `--bslib-navbar-{light,dark}-bg` variables

* feat(shiny-preset): Use border-bottom for default navbar not in a dashboard page

* refactor(shiny-preset): Use Sass vars to set navbar light/dark bg

* fix(bs5): Selector on light-region navbar-toggler-icon

* fix(navbar): Dark/light adjustments need to have the lowest specificity possible

Otherwise they get in the way of local classes, e.g. `bg-blue` or `bg-primary`,
which are the primary mechanism for setting navbar color in BS5

https://getbootstrap.com/docs/5.3/components/navbar/#color-schemes

* fix(navbar): Use `:where()` to reduce navbar dark/light styles to minimal specificity

* fix(navbar): Use `:where()` when setting navbar bg to ensure that utility classes will win

* feat!(navbars): Don't use BS3 compat for navbars in BS5+

* chore: whitespace fixes